% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickrrestquery.R
\name{flickr.clear_cache}
\alias{flickr.clear_cache}
\title{Clear cached results}
\usage{
flickr.clear_cache()
}
\description{
Clears the local cache (located at \code{.flickr_api_result} in the \code{.GlobalEnv}).
}
\examples{
flickr.clear_cache()

}
