% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickrmetadl.R
\name{flickr.meta.dl}
\alias{flickr.meta.dl}
\title{Download all metadata related to flickr.photos.search}
\usage{
flickr.meta.dl(api_key, sub_bbox, extras = "geo,tags,date_taken,url_m")
}
\arguments{
\item{api_key}{your personal API key from
\href{https://www.flickr.com/services/apps/create/apply/}{flickr}}

\item{sub_bbox}{a list of spatial bounding box from spatial data; output from
\code{splitbbox()}}

\item{extras}{character vector of potential extra fields returned. By default
('geo,tags') returns the geotag information (latitude, longitude, etc) and
the photo's tags (keywords). Currently supported fields are: description,
license, date_upload, date_taken, owner_name, icon_server, original_format,
last_update, geo, tags, machine_tags, o_dims, views, media, path_alias,
url_sq, url_t, url_s, url_q, url_m, url_n, url_z, url_c, url_l, url_o}
}
\value{
data.frame
}
\description{
Returns data.frame of metadata for geotagged images inside the bounding box for all available pages. Unavailable pages will be filled with NA
}
\examples{
api <- '1aaaaa11a1aa11aaa1a11a111a1aaaaa'   # insert your own API key here
sec <- '11a1a111a111a1a1'                   # insert your own secret code here
bb <- '-65,44.5,-64.5,45'
photos <- flickr.photos.search(api_key = api,secret = sec,bb=bb)
}

