% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickrpeoplegetInfo.R
\name{flickr.people.getInfo}
\alias{flickr.people.getInfo}
\title{Gets available user profile info}
\usage{
flickr.people.getInfo(owner)
}
\arguments{
\item{owner}{an owner value in the metadata data.frame}
}
\value{
data.frame
}
\description{
Gets available user profile info, currently only retrieve user location
}
\examples{
flickr.people.getInfo('85701104@N03')
}
