% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickrpeoplegetphotos.R
\name{flickr.people.getPhotos}
\alias{flickr.people.getPhotos}
\title{Search flickr for particular user's photos}
\usage{
flickr.people.getPhotos(api_key, user_id = NULL, extras = c("geo", "tags",
  "date_taken", "url_m"), output = "metadata", .allpages = FALSE,
  .usecache = TRUE, ...)
}
\arguments{
\item{api_key}{your personal API key from
\href{https://www.flickr.com/services/apps/create/apply/}{flickr}}

\item{user_id}{The ID of the user who's photos to return. A value of "me" will return the calling user's photos.}

\item{extras}{character vector of potential extra fields returned. By default
('geo,tags') returns the geotag information (latitude, longitude, etc) and
the photo's tags (keywords). Currently supported fields are: description,
license, date_upload, date_taken, owner_name, icon_server, original_format,
last_update, geo, tags, machine_tags, o_dims, views, media, path_alias,
url_sq, url_t, url_s, url_q, url_m, url_n, url_z, url_c, url_l, url_o}

\item{.allpages}{Pass \code{TRUE} to return all pages of results (by default
only 100 ish results are returned).}

\item{.usecache}{Pass \code{FALSE} to not use cached results.}

\item{...}{Key/value pairs as defined in
\href{https://www.flickr.com/services/api/flickr.photos.search.html}{the
API documentation}. \code{text="my search query"} is particularly useful.}
}
\value{
A \code{data.frame} of photo information.
}
\description{
An R implementation of
\href{https://www.flickr.com/services/api/explore/flickr.people.getPhotos}{flickr.people.getPhotos}
from the \href{https://www.flickr.com/services/api/}{Flickr API}. Searches for photos based on
user-defined criteria. Some parameters (\code{extras}) allow passing of R objects
that are converted to the text input needed by the API.
}
\examples{
user_id = "141233693@N02"
flickr.people.getPhotos(api_key,user_id = user_id)
}
