% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickrrestquery.R
\name{flickr.restquery}
\alias{flickr.restquery}
\title{Call the Flickr REST API}
\usage{
flickr.restquery(..., rest_api = "https://api.flickr.com/services/rest/",
  .usecache = TRUE)
}
\arguments{
\item{...}{Arguments are treated as key/value pairs to be passed to the URL.}

\item{rest_api}{The base URL for the REST API.}

\item{.usecache}{\code{TRUE} if the cache should be used, use \code{FALSE} to force a call to the
REST API.}
}
\value{
A \code{list} of the result. This list is guaranteed to have a \code{stat} and \code{result}
         element describing the result.
}
\description{
Calls the \href{Flickr REST API}{https://www.flickr.com/services/api/}, returning the parsed
JSON as a \code{list}. This is a low-level function designed to serve as the base to all
flickr method calls. Results are cached and the cache is used instead of querying the REST
API to prevent unnecessary calls (useful for repetetive scripts). To pass a \code{list} of
key/value pairs, use \code{do.call(flickr.restquery, key_value_list)}
}
\examples{
library(flickRgeotag)
flickr.restquery(method="flickr.test.echo", name="value", api_key="610ccba3a846af469e1894da33514ea1")

dta <- flickr.restquery(method="flickr.photos.search", api_key="2adee16e727679964eda05463d41fa74", bbox="-65,44.5,-64.5,45")
df <- data.frame(t(sapply(dta$photos$photo, function(item) {item})))
head(df)

}

