% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickRgeotag.R
\docType{package}
\name{flickRgeotag}
\alias{flickRgeotag}
\alias{flickRgeotag-package}
\title{flickRgeotag}
\description{
Queries flickr for geotagged data. Works as an R wrapper for the flickerapi Python module. Requires a Python installation.
}

