% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qflickrplot.R
\name{qflickr.plot}
\alias{qflickr.plot}
\title{plots flickr metadata using leaflet}
\usage{
qflickr.plot(photos)
}
\arguments{
\item{photos}{metadata data.frame from flickr.meta.dl() or flickr.photos.search()}
}
\value{
leaflet plot
}
\description{
plots flickr metadata using leaflet
}
\examples{
api <- '1aaaaa11a1aa11aaa1a11a111a1aaaaa'   # insert your own API key here
sec <- '11a1a111a111a1a1'                   # insert your own secret code here
bb <- '-65,44.5,-64.5,45'
photos <- flickr.photos.search(api_key = api,secret = sec,bbox=bb)
photosplusinfo <- qflickr.plot(photos)
}
