% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickrmetadl.R
\name{flickr.meta.dl}
\alias{flickr.meta.dl}
\title{Download all metadata related to flickr.photos.search}
\usage{
flickr.meta.dl(api_key, secret, bbox, extras)
}
\arguments{
\item{api_key}{your personal API key from \href{https://www.flickr.com/services/apps/create/apply/}{flickr}}

\item{secret}{your secret code from \href{https://www.flickr.com/services/apps/create/apply/}{flickr}}

\item{bbox}{the spatial bounding box from spatial data; output from sp::bbox() (e.g. bbox=bbox(shp)) or a character string (e.g. bbox='-65,44.5,-64.5,45')}

\item{extras}{character string of potential extra fields returned. By default ('geo,tags') returns the geotag information (latitude, longitude, etc) and the photo's tags (keywords). Currently supported fields are: description, license, date_upload, date_taken, owner_name, icon_server, original_format, last_update, geo, tags, machine_tags, o_dims, views, media, path_alias, url_sq, url_t, url_s, url_q, url_m, url_n, url_z, url_c, url_l, url_o}
}
\value{
data.frame
}
\description{
Returns data.frame of metadata for geotagged images inside the bounding box for all available pages. Unavailable pages will be filled with NA
}

