% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickrphotossearch.R
\name{flickr.photos.search}
\alias{flickr.photos.search}
\title{flickr.photos.search}
\usage{
flickr.photos.search(api_key, secret, bbox)
}
\arguments{
\item{api_key}{your personal API key from \href{https://www.flickr.com/services/apps/create/apply/}{flickr}}

\item{secret}{your secret code from \href{https://www.flickr.com/services/apps/create/apply/}{flickr}}

\item{bbox}{the spatial bounding box from spatial data; output from sp::bbox() (e.g. bbox=bbox(shp)) or a character string (e.g. bbox='-65,44.5,-64.5,45')}
}
\value{
data.frame
}
\description{
Returns data.frame of metadata for geotagged images inside the bounding box. Essentially a pared down R/Python version of the flickr \href{https://www.flickr.com/services/api/explore/flickr.photos.search}{tool}
}
\examples{
api <- '1aaaaa11a1aa11aaa1a11a111a1aaaaa'   # insert your own API key here
sec <- '11a1a111a111a1a1'                   # insert your own secret code here
bb <- '-65,44.5,-64.5,45'
photos <- flickr.photos.search(api_key = api,secret = sec,bbox=bb)
}

