% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickRgeotag.R
\docType{package}
\name{flickRgeotag}
\alias{flickRgeotag}
\alias{flickRgeotag-package}
\title{package}
\description{
Queries flickr for geotagged data. Works as a customized R wrapper for the flickr \href{https://www.flickr.com/services/api/}{API}.
}

